+++ androidApp/src/main/java/com/gdgnantes/devfest/androidapp/utils/NavigationUtils.kt
@@ -6,6 +6,7 @@ import android.location.Location
 import android.net.Uri
 
 fun Context.onNavigationClick(location: Location) {
-    val uri = "geo:latitude,longtitude?q= ${location.latitude},${location.longitude}"
+    val uri =
+        "geo:${location.latitude},${location.longitude}?q=Cité des Congrès de Nantes"
     startActivity(Intent(Intent.ACTION_VIEW, Uri.parse(uri)))
 }