+++ shared/src/commonMain/kotlin/com/gdgnantes/devfest/model/Venue.kt
@@ -3,6 +3,7 @@ package com.gdgnantes.devfest.model
 data class Venue(
     var address: String = "",
     var description: String = "",
+    var floorPlanUrl: String? = null,
     var latitude: Double? = 0.0,
     var longitude: Double? = 0.0,
     var imageUrl: String? = null,