+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 Change Log
 ==========
 
-Version 1.0.0 *(2022-09-??)*
+Version 1.0.1 *(2022-10-??)*
+----------------------------
+
+* 📣 Adds venue's floor plan.
+* 📣 Adds additional session filters.
+
+Version 1.0.0 *(2022-10-12)*
 ----------------------------
 
 * 🎉 First version.